challenge_repo/requirements.inferred.txt
# Generated by betteruv
# Review before committing for production use.

Pillow
PyJWT
PyMuPDF
PyYAML
beautifulsoup4
boto3
fastapi>=0.92.0
google
google-api-python-client
httpx>=0.23.0
numpy>=1.22.3
opencv-python
pandas>=1.4.2
pydantic>=1.10.2
pytest
python-dateutil
python-dotenv
redis>=4.2.0
requests>=2.28.1
scikit-learn
sqlalchemy>=1.4.39
uvicorn>=0.18.2
